wifi_locator_app: first successful BLE Client API Call + Unit Tests
A Class has been made to capture all the API Calls to the (BLE_GATT_Client_for_Windows.exe) application.

So far, only one function has been added that can succesfully update the BLE_Device_Name field in the Python companion app, but the JUnit tests are working.

This work will form the basis for the other functions to be added to interact with the API server.