wifi_locator_app: Added Additional OEM Lookup Functionality + GUI coordination
We have incorporated the "macManufactuerLookup" functions as part of the (Wi-Fi Scan Tab), which are invoked by pressing the corresponding button.
- This also includes implementing an additional function in "wifiScan" that modifies an input JSON string to also include the Manufacturer item.

A parallel Service has been created that periodically checks for Internet availability (which is required by some functions). This has also been integrated with a "Circle" element in the GUI.

Added additional functionality to the Controller to disable certain buttons if the conditions needed to execute their functions haven't been met.

No additional Tests have been developed for the app.