wifi_locator_app: Added ESP32 GPS Functionality + Geoapify lookup
A new set of Classes has been created for the purpose of requesting and parsing the raw GPS data that is coming from the ESP32:

- gpsCaller: Functions used to send a request to receive GPS data from ESP32
- gpSParser: Functions that can be called to parse GPRMC sentences (from the raw GPS data) into readable form.

Additionally, a new Class has been created to try to estimate the address from the provided numeric coordinates (gpsGeolocator). This function requires a user to have created an account with Geoapify and retrieved the API key for the "Reverse Geocoding API".

These changes have been integrated into the (FXML) file already.

No additional Tests have been developed for the app.