wifi_locator_app: Added Wi-Fi AP Distance Estimator Function
New functionality has been added that enables a user to try to estimate the distance from the ESP32 to a selected Wi-Fi AP based on the RSSI.

Three different values of distances are provided based on levels of attenuation that can realistically occur.

When the user wants to view the estimated distances for an AP, a new JavaFx Scene will be invoked with its own Controller (DistanceCalculatorController.java). Every time one of these scenes are invoked, it is added as a "Child" scene from the original app.

Further work should be considered to try to refine the distance estimation formula or assumptions based on inconsistent performance.

No new tests have been created.