wifi_locator_app: Added Help Menu Pop Up Scenes
New JavaFx Scenes have been added that are opened when the user selects the corresponding option in the "Help" menu of the app.

The Controllers for these Scenes are kept in Gui.HelpMenu, and they are mostly empty or perform minimal functions.

Each of these new scenes is registered as a Child Scene by calling (load_new_generic_scene), such that they will all close when the main scene is closed.

This commit also includes minor updates to function comments.

No new tests have been created.