wifi_locator_app: Added License + Launch4J Plug-Ins
Added the Maven Launch4J plugin to create an (EXE) file from this project to make it independent of needing the correct JRE installed on a platform.

Added the Maven License plugin to automatically extract a copy of most of the Maven modules that have been downloaded and used as part of this project.