wifi_locator_app: Added "Reset All" Function
The "Reset All" function can be accessed from the App Drop-Down Menu.

This resets all major Object Instances that control the state of each of the App Tabs, as well as resets the JavaFx GUI elements that are auto-populated by the App responding to the user's commands.

Note: The remaining text that is entered by the user is kept as is.