wifi_locator_app: Fixed Faulty MVN Build
There was an error where the original Maven build did not produce an executable JAR and (.EXE) build output.

I am not sure why this happened because I was able to execute this build artifact in the past, so I have not changed the build to use the (maven-assembly-plugin) instead of the (maven-jar-plugin).